Revert "Sonar: mitigate inheritance abuse."
This reverts commit ae0d53370c338030c4a2741b0aa93aaf6dab8e94.
