+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/AbstractBlackBoxModel.java
@@ -53,11 +53,6 @@ public abstract class AbstractBlackBoxModel implements BlackBoxModel, DynamicMod
     }
 
     @Override
-    public void writeMacroConnect(XMLStreamWriter writer, DynaWaltzContext context, MacroConnector macroConnector, BlackBoxModel connected) throws XMLStreamException {
-        macroConnector.writeMacroConnect(writer, List.of(Pair.of("id1", getDynamicModelId())), connected.getAttributesConnectTo());
-    }
-
-    @Override
     public List<Pair<String, String>> getVarsConnect(BlackBoxModel connectedBbm) {
         return Collections.emptyList();
     }
