+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/nonstaticref/automatons/AbstractAutomaton.java
@@ -1,26 +1,17 @@
 package com.powsybl.dynawaltz.dynamicmodels.nonstaticref.automatons;
 
 import com.powsybl.dynawaltz.DynaWaltzContext;
-import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModel;
-import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModelWithStaticId;
+import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModelWithStaticId;
 
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
 
 import static com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants.DYN_URI;
 
-public abstract class AbstractAutomaton extends AbstractBlackBoxModel implements BlackBoxModelWithStaticId {
-
-    private final String staticId;
+public abstract class AbstractAutomaton extends AbstractBlackBoxModelWithStaticId {
 
     protected AbstractAutomaton(String dynamicModelId, String staticId, String parametersId) {
-        super(dynamicModelId, parametersId);
-        this.staticId = staticId;
-    }
-
-    @Override
-    public String getStaticId() {
-        return staticId;
+        super(dynamicModelId, staticId, parametersId);
     }
 
     protected void writeAutomatonBlackBoxModel(XMLStreamWriter writer, DynaWaltzContext context) throws XMLStreamException {
