+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/nonstaticref/events/AbstractBlackBoxEventModel.java
@@ -8,7 +8,7 @@ package com.powsybl.dynawaltz.dynamicmodels.nonstaticref.events;
 
 import com.powsybl.dynamicsimulation.EventModel;
 import com.powsybl.dynawaltz.DynaWaltzContext;
-import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModel;
+import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModelWithStaticId;
 import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModel;
 import com.powsybl.dynawaltz.dynamicmodels.MacroConnector;
 import org.apache.commons.lang3.tuple.Pair;
@@ -22,18 +22,10 @@ import static com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants.DYN_URI;
 /**
  * @author Marcos de Miguel <demiguelm at aia.es>
  */
-public abstract class AbstractBlackBoxEventModel extends AbstractBlackBoxModel implements EventModel, BlackBoxEventModel {
-
-    private final String staticId;
+public abstract class AbstractBlackBoxEventModel extends AbstractBlackBoxModelWithStaticId implements EventModel, BlackBoxEventModel {
 
     protected AbstractBlackBoxEventModel(String eventModelId, String staticId, String parameterSetId) {
-        super(eventModelId, parameterSetId);
-        this.staticId = staticId;
-    }
-
-    @Override
-    public String getStaticId() {
-        return staticId;
+        super(eventModelId, staticId, parameterSetId);
     }
 
     @Override
