+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/nonstaticref/network/AbstractNetworkBlackBoxModel.java
@@ -1,24 +1,15 @@
 package com.powsybl.dynawaltz.dynamicmodels.nonstaticref.network;
 
-import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModel;
-import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModelWithStaticId;
+import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModelWithStaticId;
 import com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.List;
 
-public abstract class AbstractNetworkBlackBoxModel extends AbstractBlackBoxModel implements BlackBoxModelWithStaticId {
-
-    private final String staticId;
+public abstract class AbstractNetworkBlackBoxModel extends AbstractBlackBoxModelWithStaticId {
 
     protected AbstractNetworkBlackBoxModel(String staticId) {
-        super("", "");
-        this.staticId = staticId;
-    }
-
-    @Override
-    public String getStaticId() {
-        return staticId;
+        super("", staticId, "");
     }
 
     @Override
