+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/staticref/AbstractBlackBoxModelWithStaticRef.java
@@ -7,8 +7,7 @@
 package com.powsybl.dynawaltz.dynamicmodels.staticref;
 
 import com.powsybl.dynawaltz.DynaWaltzContext;
-import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModel;
-import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModelWithStaticId;
+import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModelWithStaticId;
 import com.powsybl.dynawaltz.xml.MacroStaticReference;
 import org.apache.commons.lang3.tuple.Pair;
 
@@ -22,20 +21,13 @@ import static com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants.DYN_URI;
 /**
  * @author Luma Zamarreño <zamarrenolm at aia.es>
  */
-public abstract class AbstractBlackBoxModelWithStaticRef extends AbstractBlackBoxModel implements BlackBoxModelWithStaticId, BlackBoxModelWithStaticRef {
+public abstract class AbstractBlackBoxModelWithStaticRef extends AbstractBlackBoxModelWithStaticId implements BlackBoxModelWithStaticRef {
 
-    private final String staticId;
     private final List<Pair<String, String>> staticRef;
 
     protected AbstractBlackBoxModelWithStaticRef(String dynamicModelId, String staticId, String parameterSetId, List<Pair<String, String>> staticRef) {
-        super(dynamicModelId, parameterSetId);
-        this.staticId = staticId;
+        super(dynamicModelId, staticId, parameterSetId);
         this.staticRef = staticRef;
-    }
-
-    @Override
-    public String getStaticId() {
-        return staticId;
     }
 
     @Override
